layed out the look of my sidebar buttons
